requirements.txt
BnLemma==2.1
BnLemma==2.1
Django==5.0.3
djangorestframework==3.15.0
nltk==3.8.1
numpy==1.26.4
pandas==2.2.2
scikit_learn==1.4.1.post1
scipy==1.13.0
sentence_transformers==2.6.1
torch==2.2.2
transformers==4.39.3
